+++ src/alfrescoApiClient.js
@@ -62,6 +62,11 @@ class AlfrescoApiClient extends ApiClient {
             this.setCsrfToken(request);
         }
 
+        // add cookie for activiti
+        if (this.isBpmRequest() && this.cookie) {
+            request.set('Cookie', this.cookie);
+        }
+
         // set request timeout
         request.timeout(this.timeout);
 
@@ -120,6 +125,11 @@ class AlfrescoApiClient extends ApiClient {
                     }
 
                 } else {
+                    if (this.isBpmRequest()) {
+                        if (response.header && response.header.hasOwnProperty('set-cookie')) {
+                            this.cookie = response.header['set-cookie'];
+                        }
+                    }
                     var data = {};
                     if (response.type === 'text/html') {
                         data = this.deserialize(response, 'String');